appreasy/backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy[asyncio]
psycopg[binary]
geoalchemy2
pydantic
python-dotenv
requests
pandas
asyncpg
APScheduler
pyjwt[crypto]
python-multipart
Shapely
mapbox-vector-tile
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
aiohttp
pytest
